requirements.txt
requests>=2.31.0
matplotlib>=3.7.0
numpy>=1.24.0
python-dotenv>=1.0.0
